+++ zakozhurnikova.kristina/P4/main.cpp
@@ -5,7 +5,7 @@
 
 int main(int argc, char * argv[])
 {
-  int num = 0;
+  int option = 0;
   try
   {
     zakozhurnikova::rightArguments(argc, argv);
@@ -19,6 +19,7 @@ int main(int argc, char * argv[])
   size_t cols = 0;
   std::ifstream input(argv[2]);
   input >> rows >> cols;
+  const int s = rows * cols;
   if (!input)
   {
     std::cerr << "Cannot read a number\n";
@@ -38,8 +39,26 @@ int main(int argc, char * argv[])
       input.close();
       return 2;
     }
-    input.close();
-    //cod1
+    int matrix[s]{};
+    int original[s]{};
+    fillMatrix(matrix, rows, cols);
+
+    for (int i = 0; i < rows; i++)
+    {
+      for (int j = 0; j < cols; j++)
+      {
+        input >> original[i*cols + j];
+      }
+    }
+    substractMatrix(original, matrix, rows, cols);
+    for (int i = 0; i < rows; i++)
+    {
+      for (int j = 0; j < cols; j++)
+      {
+        std::cout << original[i*rows + j] << ' ';
+      }
+      std::cout << '\n';
+    }
   }
   else
   {
@@ -64,48 +83,5 @@ int main(int argc, char * argv[])
     input.close();
   }
   return 0;
-}
-
-//namespace.cpp
-int zakozhurnikova::rightArguments(int argc, char** argv)
-{
-  if (argc != 4)
-  {
-    throw std::invalid_argument("Error in command line arguments\n");
-    return 1;
-  }
-  char* ptr = nullptr;
-  int num = std::strtol(argv[1], &ptr, 10);
-  if (num == 0 && ptr == argv[1])
-  {
-    throw std::invalid_argument("First parameter is not a number");
-    return 1;
-  }
-  if (num != 2 && num != 1)
-  {
-    throw std::invalid_argument("First parameter is not 1 and 2");
-    return 1;
-  }
-  return num;
-}
-void zakozhurnikova::inputMatrix(std::ifstream& input, double* size, size_t rows, size_t cols)
-{
-  size_t account = 0;
-  for (size_t i = 0; i < rows * cols; ++i)
-  {
-    input >> *size;
-    ++size;
-    if (input)
-    {
-      ++account;
-    }
-  }
-  char check_size = '\0';
-  input >> check_size;
-  if (account != rows * cols || input)
-  {
-    throw std::range_error("Input data is not desired matrix");
-  }
-  return;
 }
